crucible: fix crash and filter sentences
Bugfix: report.validator causes exceptions on invalid confidences, this
led to queries being dropped. This is fixed now. Invalid confidences
will be set to the right range. A warning is triggered.  Other
exceptions will halt the whole program.

We discourage sentences that contain certain expressions like "source
document" which indicate that a "non-extraction" was made such as "the
source document does not contain the answer to this question". Such
sentences will get the confidence halved.